Start reading response before request sent. Added retry for init poll.
There was a bug where the device request was sent, but the response was
delivered when we did not yet start to read the response, thus losing
the response and ending in timeout. Now the with MessageReader, we start
to read the response before the request is sent.

Added a retry pattern on simple poll on initialisation to avoid error
caused by simple timeout when there is no real issue.

Simplified devices and handlers exception tree.