Abstracted SerialPort implementation. Added device configurator.
The SerialPort ccTalk4j interface is used to abstract away the concrete
Java implementation (currently jssc). This allow to have an access to
serial port configuration without worrying about Serial implementation
details, and using a different implementation on different platforms.

This was done because of Serial issues between Windows and Linux, most
Java serial implementations behaving differently on each platform, and
the Java Serial libraries available at the time are not quite
satisfying. 

The device configurator pattern is used to easily configure the device
before connecting it - depending on serial implementation, it may be
required to configure the port AFTER having it connected.